Move NextBillingPeriodStartDate to holid-stops library under subscriptions package
