+++ handlers/holiday-stop-processor/src/test/scala/com/gu/holidaystopprocessor/HolidayCreditUpdateTest.scala
@@ -2,6 +2,7 @@ package com.gu.holidaystopprocessor
 
 import java.time.LocalDate
 
+import com.gu.holiday_stops
 import com.gu.holiday_stops.Fixtures.guardianWeeklyConfig
 import com.gu.holiday_stops._
 import com.gu.holiday_stops.subscription._
@@ -20,7 +21,7 @@ class HolidayCreditUpdateTest extends FlatSpec with Matchers with EitherValues {
       chargedThroughDate = Some(LocalDate.of(2019, 9, 12))
     )
     val currentGuardianWeeklySubscription = CurrentGuardianWeeklySubscription(subscription, Fixtures.config)
-    val nextInvoiceStartDate = NextBillingPeriodStartDate(Fixtures.config, subscription, LocalDate.now).right.value
+    val nextInvoiceStartDate = holiday_stops.subscription.NextBillingPeriodStartDate(Fixtures.config, subscription, LocalDate.now).right.value
     val maybeExtendedTerm = ExtendedTerm(nextInvoiceStartDate, subscription)
     val holidayCredit = GuardianWeeklyHolidayCredit(currentGuardianWeeklySubscription.right.value, LocalDate.now)
 
@@ -74,7 +75,7 @@ class HolidayCreditUpdateTest extends FlatSpec with Matchers with EitherValues {
       chargedThroughDate = Some(LocalDate.of(2020, 8, 2))
     )
     val currentGuardianWeeklySubscription = CurrentGuardianWeeklySubscription(subscription, Fixtures.config)
-    val nextInvoiceStartDate = NextBillingPeriodStartDate(Fixtures.config, subscription, LocalDate.now).right.value
+    val nextInvoiceStartDate = holiday_stops.subscription.NextBillingPeriodStartDate(Fixtures.config, subscription, LocalDate.now).right.value
     val maybeExtendedTerm = ExtendedTerm(nextInvoiceStartDate, subscription)
     val holidayCredit = GuardianWeeklyHolidayCredit(currentGuardianWeeklySubscription.right.value, LocalDate.now)
     val update = HolidayCreditUpdate(
@@ -114,7 +115,7 @@ class HolidayCreditUpdateTest extends FlatSpec with Matchers with EitherValues {
       chargedThroughDate = Some(LocalDate.of(2020, 7, 23))
     )
     val currentGuardianWeeklySubscription = CurrentGuardianWeeklySubscription(subscription, Fixtures.config)
-    val nextInvoiceStartDate = NextBillingPeriodStartDate(Fixtures.config, subscription, LocalDate.now).right.value
+    val nextInvoiceStartDate = holiday_stops.subscription.NextBillingPeriodStartDate(Fixtures.config, subscription, LocalDate.now).right.value
     val maybeExtendedTerm = ExtendedTerm(nextInvoiceStartDate, subscription)
     val holidayCredit = GuardianWeeklyHolidayCredit(currentGuardianWeeklySubscription.right.value, LocalDate.now)
     val update = HolidayCreditUpdate(
